requirements.txt
Django==3.2.3
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
pytest==3.6.0
pytest-django==3.2.1
django-debug-toolbar==1.9.1
djangorestframework
pyjwt==1.7.0
django-cors-headers